Rewrite with improved prompting and code
I wrote this almost exactly as I would write it myself; hope it helps.

* Added my usual Err() function for easy error messaging
* Added various error messages regarding the prompts
* Changed shebang to more portable version
* Tweaked prompt output to be more concise
* Improved prompt code to be clearer and easier to work with
* Added ability to just quit from the script
* Added dependency check for `docker` in PATH

This PR was sent following an E-Mail by you, in-case you ever look back
at this in the future and wonder why this came to be. If you have any
questions or requests, let me know.

Be sure to _test_ this _before_ pulling, although it should be fine, -
but better safe than sorry.

The reason I changed "The following script" to "This script" is because
the former suggests there's another to come. The latter is more
accurate. Don't mind me, I'm just a bit pedantic.

If you absolutely must echo out the Docker command first, be sure to
keep it separate so it's clear what is being echoed and what is being
executed.

Try to avoid confusing quotations, as it makes it harder to read.