Add check for valid version string (#2)
From looking here...

https://github.com/portainer/portainer/releases?page=1

...I've determined what a valid Portainer version string is. If they
ever change it, then you'll have to update the REGEX. They've used the
same string for ages though, so it'll probably be fine.

You _could_ add a check (to the function) for whether the string you
enter is the same as the one you have, but I don't know how to get the
existing Portainer version. Even if I looked that up, I can't guarantee
I'd be parsing the right thing, because I'm not a Docker or Portainer
user.